refactor(agent): phase C — decompose telegram.rs into client/formatting/templates/burst
